Move classes from ExecutionEngine to separate files
