+++ Source/ExecutionEngine/PipelineOutcome.cs
@@ -1,0 +1,12 @@
+namespace Microsoft.Boogie;
+
+public enum PipelineOutcome
+{
+  Done,
+  ResolutionError,
+  TypeCheckingError,
+  ResolvedAndTypeChecked,
+  FatalError,
+  Cancelled,
+  VerificationCompleted
+}
